requirements.txt
flask
mysql-connector-python
flask-sqlalchemy
flask-WTF
flask-Bcrypt
